+++ src/battle_transition.c
@@ -3814,7 +3814,7 @@ void Task_BattleTransition_Intro(u8 taskId)
 
 static bool8 TransitionIntro_FadeToGray(struct Task *task)
 {
-    u8 paletteNum = IndexOfSpritePaletteTag(TAG_WEATHER_START);
+    u32 paletteNum = IndexOfSpritePaletteTag(TAG_WEATHER_START);
     u16 index = OBJ_PLTT_ID2(paletteNum) + SHADOW_COLOR_INDEX;
     if (task->tDelayTimer == 0 || --task->tDelayTimer == 0)
     {
@@ -3847,7 +3847,7 @@ static bool8 TransitionIntro_FadeFromGray(struct Task *task)
 {
     if (task->tDelayTimer == 0 || --task->tDelayTimer == 0)
     {
-        u8 paletteNum = IndexOfSpritePaletteTag(TAG_WEATHER_START);
+        u32 paletteNum = IndexOfSpritePaletteTag(TAG_WEATHER_START);
         task->tDelayTimer = task->tFadeFromGrayDelay;
         task->tBlend -= task->tFadeFromGrayIncrement;
         if (task->tBlend < 0)
